# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Flyspray",
        "version": "1.0.0",
        "contact": {}
    },
    "paths": {
...
        "/comment": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Creates a new comment for a bug in the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comments"
                ],
                "summary": "Creates a comment",
                "parameters": [
                    {
                        "description": "Comment input object",
                        "name": "commentInput",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/app.createCommentInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Comment is created successfully (Comment details in the 'Data' field)",
                        "schema": {
                            "$ref": "#/definitions/app.ResponseMsg"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/comment/filters": {
            "get": {
                "description": "Retrieves a list of comments from the database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comments"
                ],
                "summary": "Retrieves comments",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Bug ID",
                        "name": "bug_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "comments are retrieved successfully (Comments details in the 'Data' field)",
                        "schema": {
                            "$ref": "#/definitions/app.ResponseMsg"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/comment/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieves a comment by its ID from the database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comments"
                ],
                "summary": "Retrieves a comment",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Comment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "comment is found successfully (Comment details in the 'Data' field)",
                        "schema": {
                            "$ref": "#/definitions/app.ResponseMsg"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Updates a comment by its ID in the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comments"
                ],
                "summary": "Updates a comment",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Comment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated comment data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/app.updateCommentInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "comment is updated successfully",
                        "schema": {
                            "$ref": "#/definitions/app.ResponseMsg"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {}
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Deletes a comment by its ID from the database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comments"
                ],
                "summary": "Deletes a comment",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Comment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "comment is deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/app.ResponseMsg"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {}
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
...
    },
    "definitions": {
        "app.ResponseMsg": {
            "type": "object",
            "properties": {
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "app.createCommentInput": {
            "type": "object",
            "required": [
                "bug_id",
                "summary"
            ],
            "properties": {
                "bug_id": {
                    "type": "integer"
                },
                "summary": {
                    "type": "string"
                }
            }
        },
...
        "app.updateCommentInput": {
            "type": "object",
            "required": [
                "summary"
            ],
            "properties": {
                "summary": {
                    "type": "string"
                }
            }
        },
...
    }
}